backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-genai
requests
selenium
beautifulsoup4
lxml
numpy
pandas
scikit-learn
